[Fix #702] Revert adding domain and range to properties
