+++ src/main/java/cz/cvut/kbss/termit/model/CustomAttribute.java
@@ -5,6 +5,7 @@ import cz.cvut.kbss.jopa.model.annotations.Context;
 import cz.cvut.kbss.jopa.model.annotations.OWLAnnotationProperty;
 import cz.cvut.kbss.jopa.model.annotations.OWLClass;
 import cz.cvut.kbss.jopa.vocabulary.RDF;
+import cz.cvut.kbss.jopa.vocabulary.RDFS;
 import cz.cvut.kbss.termit.util.Vocabulary;
 
 import java.net.URI;
@@ -15,6 +16,12 @@ import java.util.Set;
 public class CustomAttribute extends RdfsResource {
 
     public static final String CONTEXT = Vocabulary.s_i_termit + "/custom-attributes";
+
+    @OWLAnnotationProperty(iri = RDFS.DOMAIN)
+    private URI domain;
+
+    @OWLAnnotationProperty(iri = RDFS.RANGE)
+    private URI range;
 
     @OWLAnnotationProperty(iri = Vocabulary.s_p_as_relationship)
     private Set<URI> annotatedRelationships;
@@ -27,6 +34,21 @@ public class CustomAttribute extends RdfsResource {
         super(uri, label, comment, RDF.PROPERTY);
     }
 
+    public URI getDomain() {
+        return domain;
+    }
+
+    public void setDomain(URI domain) {
+        this.domain = domain;
+    }
+
+    public URI getRange() {
+        return range;
+    }
+
+    public void setRange(URI range) {
+        this.range = range;
+    }
 
     public Set<URI> getAnnotatedRelationships() {
         return annotatedRelationships;
